Rename imagenet_dataset_np => imagenet_dataset
