+++ dl_playground/datasets/imagenet_dataset.py
