+++ tests/integration_tests/datasets/test_tf_data_loader.py
@@ -3,7 +3,7 @@
 import numpy as np
 import tensorflow as tf
 
-from datasets.imagenet_dataset_np import ImageNetDataSet
+from datasets.imagenet_dataset import ImageNetDataSet
 from datasets.ops import resize_images
 from datasets.tf_data_loader import TFDataLoader
 from utils.test_utils import df_images
